refactor(backend): Simplify payload user_id retrieval and refactor te...
- Remove redundant `payload` existence check in `get_user_preferences`
- Extract shared test stubs (`RequestStub`, `ResultStub`, `DbStub`) into `tests/stubs.py`
- Update `test_language_switch_flow.py` and `test_webui_user_preferences.py` to import stubs from `tests.stubs`
- Add `@pytest.fixture` to clear `_USER_PREFS_CACHE` before and after each test
- Simplify assertions in `test_language_switch_full_flow` by removing intermediate variable
- Update docstring in `test_language_switch_flow.py` to be more concise
